+++ src/OpenCOVER/cover/mui/support/coMUIElementManager.cpp
@@ -24,7 +24,7 @@ struct coMUIElementManager::entry{
 // adds container-element to ElementList
 void coMUIElementManager::addElement(std::string UniqueIdentifier, coMUIContainer* Container){
     bool exist=0;
-    for (int i=0; i<ElementList.size(); i++){
+    for (size_t i=0; i<ElementList.size(); i++){
         if (ElementList[i].name==UniqueIdentifier){                     // name already exists
             exist=1;
             std::cerr << "ERROR: coMUIElementManager::addElement(): Element named " << ElementList[i].name << " already exists. Choose another Name" << std::endl;
@@ -41,7 +41,7 @@ void coMUIElementManager::addElement(std::string UniqueIdentifier, coMUIContaine
 // adds widget-element to ElementList
 void coMUIElementManager::addElement(std::string UniqueIdentifier, coMUIWidget* Widget){
     bool exist=0;
-    for (int i=0; i<ElementList.size(); i++){
+    for (size_t i=0; i<ElementList.size(); i++){
         if (ElementList[i].name==UniqueIdentifier){                     // name already exists
             exist=1;
             std::cerr << "ERROR: coMUIElementManager::addElement(): Element named " << ElementList[i].name << " already exists. Choose another Name" << std::endl;
@@ -57,7 +57,7 @@ void coMUIElementManager::addElement(std::string UniqueIdentifier, coMUIWidget* 
 
 // removes Element from ElementList
 void coMUIElementManager::removeElement(std::string UniqueIdentifier){
-    for (int i=0; i<ElementList.size(); i++){
+    for (size_t i=0; i<ElementList.size(); i++){
         if (ElementList[i].name== UniqueIdentifier){
             ElementList.erase(ElementList.begin()+i);
             --i;
@@ -68,7 +68,7 @@ void coMUIElementManager::removeElement(std::string UniqueIdentifier){
 // prints all names from ElementList to console
 void coMUIElementManager::printNames(){
     std::string toPrint = " ";
-    for (int i=0; i< ElementList.size(); ++i){
+    for (size_t i=0; i< ElementList.size(); ++i){
         toPrint.append(ElementList[i].name);
         toPrint.append(", ");
     }
@@ -77,7 +77,7 @@ void coMUIElementManager::printNames(){
 
 // returns the container named "Identifier"
 coMUIContainer* coMUIElementManager::getContainerByIdentifier(std::string UniqueIdentifier){
-    for (int i=0; i<ElementList.size(); i++){                           // go through all entrys
+    for (size_t i=0; i<ElementList.size(); i++){                           // go through all entrys
         if ((ElementList[i].name==UniqueIdentifier) && (ElementList[i].Container == true)){                       // match (name is equal)
             return ElementList[i].ContainerPointer;
         }
@@ -88,7 +88,7 @@ coMUIContainer* coMUIElementManager::getContainerByIdentifier(std::string Unique
 
 // returns the widget named "Identifier"
 coMUIWidget* coMUIElementManager::getWidgetByIdentifier(std::string UniqueIdentifier){
-    for (int i=0; i<ElementList.size(); i++){                // go through all entrys
+    for (size_t i=0; i<ElementList.size(); i++){                // go through all entrys
         if ((ElementList[i].name==UniqueIdentifier) && (ElementList[i].Container == false)){                       // match (name is equal)
             return ElementList[i].WidgetPointer;
         }
@@ -99,7 +99,7 @@ coMUIWidget* coMUIElementManager::getWidgetByIdentifier(std::string UniqueIdenti
 
 // returns true, if element is a container; else returns false
 bool coMUIElementManager::isContainer(const std::string UniqueIdentifier){
-    for (int i=0; i<ElementList.size(); i++){
+    for (size_t i=0; i<ElementList.size(); i++){
         if (ElementList[i].name == UniqueIdentifier){
             return ElementList[i].Container;
         }
@@ -111,7 +111,7 @@ bool coMUIElementManager::isContainer(const std::string UniqueIdentifier){
 
 // delete entry with name "Name" from  ElementList
 void coMUIElementManager::deleteEntry(std::string UniqueIdentifier){
-    for (int i=0; i<ElementList.size(); i++){
+    for (size_t i=0; i<ElementList.size(); i++){
         if (ElementList[i].name == UniqueIdentifier){
             ElementList.erase(ElementList.begin()+i);
             i--;
